New CRUD Mascotas

+++ src/app/pages/mascotas/mascotas.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import {MessageService, SelectItem } from 'primeng/api';
+import {ConfirmationService, MessageService, SelectItem } from 'primeng/api';
 import { Mascota } from '../../model/mascota';
 import { MascotaService } from '../../services/mascota.service';
 import { Table } from 'primeng/table'
@@ -8,14 +8,17 @@ import { Table } from 'primeng/table'
   selector: 'app-mascotas',
   templateUrl: './mascotas.component.html',
   styleUrls: ['./mascotas.component.css', '../../app.component.css'],
-  providers: [MessageService]
+  providers: [MessageService, ConfirmationService]
 })
 export class MascotasComponent implements OnInit {
   newMascotaDialog: boolean = false;
+  viewMascotaDialog: boolean = false;
 
   mascotas: Mascota[] = [];
 
   mascota!: Mascota;
+
+  selectedMascota!: Mascota;
 
   submitted: boolean = false;
 
@@ -23,7 +26,7 @@ export class MascotasComponent implements OnInit {
 
   clonedMascotas: { [s: string]: Mascota } = {};
 
-  constructor(private mascotaService: MascotaService, private messageService: MessageService) {}
+  constructor(private mascotaService: MascotaService, private messageService: MessageService, private confirmationService: ConfirmationService) {}
 
   ngOnInit(): void {
     this.mascotaService.getMascotas().subscribe(
@@ -36,10 +39,18 @@ export class MascotasComponent implements OnInit {
     ];
   }
 
-  deleteMascota(id: number): void {
-    this.mascotaService.deleteMascota(id);
-    //Agregar nuevamente validación para eliminar mascota de la lista
-    this.mascotas = this.mascotas.filter(mascota => mascota.id !== id);
+  deleteMascota(id: number, nombre: string): void {
+    this.confirmationService.confirm({
+      message: '¿Estas seguro que quieres eliminar a ' + nombre + '?',
+      header: 'Confirmar eliminación',
+      icon: 'pi pi-exclamation-triangle',
+      accept: () => {
+        this.mascotaService.deleteMascota(id);
+        //Agregar nuevamente validación para eliminar mascota de la lista
+        this.mascotas = this.mascotas.filter(mascota => mascota.id !== id);
+        this.messageService.add({ severity: 'success', summary: '¡Exitoso!', detail: 'Mascota eliminada', life: 3000 });
+      }
+    });
   }
 
   getSeverity(status: string) {
@@ -60,7 +71,7 @@ export class MascotasComponent implements OnInit {
   onRowEditSave(mascota: Mascota) {
     //PENDIENTE: Agregar validaciones para actualizar mascota
     this.mascotaService.updateMascota(mascota);
-    this.messageService.add({ severity: 'success', summary: 'Success', detail: 'Mascota actualizada' });
+    this.messageService.add({ severity: 'success', summary: '¡Exitoso!', detail: 'Mascota actualizada' });
   }
 
   onRowEditCancel(mascota: Mascota, index: number) {
@@ -74,11 +85,33 @@ export class MascotasComponent implements OnInit {
     this.mascotasTable?.filterGlobal(($event.target as HTMLInputElement).value, stringVal);
   }
 
+  //Para ver detalles de mascota
+  openView(mascota: Mascota) {
+    this.selectedMascota = mascota;
+    this.viewMascotaDialog = true;
+  }
+
   //Para crear una nueva mascota
   openNew() {
-    this.mascota = {id: 0, nombre: '', edad: 0, raza: '', peso: 0, enfermedad: '', estado: '', fechaEntrada: '', fechaSalida: '', medicamento: '', foto: ''};
+    this.mascota = {nombre: '', edad: 0, raza: '', peso: 0, enfermedad: '', estado: '', fechaEntrada: '', fechaSalida: '', medicamento: '', foto: ''};
     this.submitted = false;
     this.newMascotaDialog = true;
+  }
+
+  hideDialog() {
+    this.newMascotaDialog = false;
+    this.submitted = false;
+  }
+
+  saveMascota() {
+    this.submitted = true;
+
+    // Imagen por defecto
+    // this.product.image = 'product-placeholder.svg';
+    this.mascotaService.addMascota(this.mascota);
+    this.messageService.add({ severity: 'success', summary: '¡Exitoso!', detail: 'Mascota creada', life: 3000 });
+    this.mascotas.push(this.mascota);
+    this.newMascotaDialog = false;
   }
 
   exportExcel() {
